de-monolith Stage 4: relocate SG2 containment to kirra-core; make the planner lean
Move the SG2 drivable-space containment checker (`VehicleFootprint` / `Corridor` /
`Pose` / `validate_trajectory_containment` / `MAX_TRAJECTORY_HORIZON`) verbatim from
`src/gateway/containment.rs` into the lean `kirra-core` crate, alongside the
kinematics-contract talisman moved in Stage 3. The relocation is byte-identical except
the file-path header comment and the intra-crate import path
(`crate::gateway::kinematics_contract` -> `crate::kinematics_contract`); all 760 logic
lines and the 19 unit tests move unchanged. The gateway module becomes a thin
`pub use kirra_core::containment::*;` re-export, so every existing
`kirra_runtime_sdk::gateway::containment::*` path (the ROS2 adapter, the planner) keeps
the SAME type — zero churn, zero logic change.

With both `FleetPosture` and the SG2 containment cap now in `kirra-core`, the planner's
own code no longer needs the heavy verifier-service tree: its direct `kirra-runtime-sdk`
dependency is replaced with `kirra-core`. (The adapter dependency remains until the
adapter is de-monolithed in a later stage, so the planner still pulls the SDK
transitively; this stage removes only the direct edge.) All `FleetPosture` imports
across the planner lib, mick module, two integration tests, and the example are
repointed to `kirra_core`.

The isolated `kirra-governor-service` already depended on `kirra-core` directly after
Stage 3; its now-stale comments describing the removed `#[path]` include are tidied to
describe the real `kirra-core` dependency.

kirra-core mirrors the parent crate's `#![allow(clippy::doc_lazy_continuation,
clippy::doc_overindented_list_items)]` so the verbatim safety-derivation doc tables keep
their byte-identical comments; serde_json is added as a dev-dependency for the moved
containment tests (test-only — the shipping crate stays serde-only).

Verification: full root workspace builds and tests green (kirra-core 61 tests incl. the
moved containment suite; planner 102 + integration/example green); CI-equivalent
`cargo clippy --workspace -- -D warnings` clean; parko workspace logic tests green
(parko-core 204, parko-kirra 148); no remaining `#[path]` include hazards.
