+++ crates/kirra-governor-service/Cargo.toml
@@ -4,11 +4,12 @@
 # prototype (docs/adr/KIRRA_BRINGUP_RUNBOOK.md, Prompt A).
 #
 # DEPENDENCY DISCIPLINE (ADR-0001): this target depends ONLY on serde + bincode
-# + std. It deliberately does NOT depend on the parent `kirra-runtime-sdk` crate
-# (which pulls tokio/axum/rusqlite/reqwest) — instead it compiles the EXISTING
-# verdict core (src/gateway/kinematics_contract.rs) in verbatim via `#[path]`
-# (see src/main.rs), so the verdict logic is the real, unmodified one while the
-# dependency tree stays minimal and ROS/async-free for the QNX cert target.
+# + std (via the lean `kirra-core` crate). It deliberately does NOT depend on the
+# parent `kirra-runtime-sdk` crate (which pulls tokio/axum/rusqlite/reqwest) —
+# instead it depends on `kirra-core`, which carries the EXISTING verdict core (the
+# kinematics-contract talisman) verbatim, so the verdict logic is the real,
+# unmodified one while the dependency tree stays minimal and ROS/async-free for the
+# QNX cert target.
 [package]
 name = "kirra-governor-service"
 version = "0.1.0"
